Mudança na logica de personagem e da logica do bot, como shutdown e aplicativo.

+++ index.js
@@ -1,4 +1,5 @@
 const { Client, GatewayIntentBits, SlashCommandBuilder, EmbedBuilder } = require('discord.js');
+const path = require('path');
 
 const client = new Client({
     intents: [
@@ -14,15 +15,25 @@ client.once('ready', async () => {
     try {
         const ping = await guild.commands.create(
             new SlashCommandBuilder()
-                .setName('ping')
-                .setDescription('Responde com Pong! Exibe informações adicionais')
+                .setName('ku')
+                .setDescription('Exibe informações do personagem')
+                .addStringOption(option =>
+                    option.setName('nome')
+                        .setDescription('Nome do personagem')
+                        .setRequired(true)
+                )
         );
         const shutdown = await guild.commands.create(
             new SlashCommandBuilder()
                 .setName('shutdown')
                 .setDescription('Desativa o bot')
+                .addStringOption(option =>
+                    option.setName('senha')
+                        .setDescription('Senha para poder desligar o bot')
+                        .setRequired(true)
+                )
         );
-        console.log('Comando /ping registrado no servidor!');
+        console.log('Comandos registrados no servidor!');
     } catch (error) {
         console.error('Erro ao registrar o comando:', error);
     }
@@ -32,21 +43,38 @@ client.on('interactionCreate', async (interaction) => {
     if (!interaction.isCommand()) return;
 
     if (interaction.commandName === 'shutdown') {
-        client.destroy();
+        const senha = interaction.options.getString('senha');
+
+        if (senha === '123456') {
+            await interaction.reply('Bot desligado!');
+            client.destroy();
+        } else {
+            await interaction.reply('Senha errada!');
+        }
     }
-    if (interaction.commandName === 'ping') {
+    if (interaction.commandName === 'ku') {
+        const nome = interaction.options.getString('nome');
+        let personagem;
+
+        try {
+            personagem = require(path.resolve(__dirname, 'personagens', `${nome}.js`));
+        } catch (error) {
+            return interaction.reply({ content: 'Personagem não encontrado!', ephemeral: true });
+        }
 
         const pingEmbed = new EmbedBuilder()
             .setColor('#33f1ff')
-            .setTitle('Ressonadors')
+            .setTitle(personagem.name)
             .addFields(
-                { name: 'Ressonadora', value: 'Shorekeep <:Rectifier:1305619439705788531> <:spectro:1305619259875135508>', inline: true },
-                { name: 'Raridade', value: '⭐⭐⭐⭐⭐', inline: true },
-                { name: 'Localização', value: '\u200B<:Black_Shores_Emblem:1309710724284551178>\u200B', inline: true },
-                
+                { name: `Raridade`, value: personagem.rarity, inline: false },
             )
-            .setThumbnail('https://cdn.discordapp.com/attachments/1055633011544170562/1310800302433112094/55.png?ex=67468969&is=674537e9&hm=c52b68e8ede1d7d3f64cfb9c3219a2615eed6ba5d2402b3dbe4d448837755e93&') // Imagem pequena à direita
-            .setImage('https://cdn.discordapp.com/attachments/1273496005484216390/1310758140953759744/Black_Shores_Emblem.webp?ex=67466225&is=674510a5&hm=542589e6591749b78002084275e1c074df613f12fc64509101ba2afac8365a0b&') // Imagem grande na parte inferior
+            .addFields(
+                { name: `Elemento`, value: personagem.element, inline: true },
+                { name: `Arma`, value: personagem.gun, inline: true },
+                { name: `Localização`, value: personagem.location, inline: true },
+            )
+            .setThumbnail(personagem.miniimage)
+            .setImage(personagem.image)
             .setFooter({
                 text: 'Informações sobre o bot',
                 iconURL: client.user.displayAvatarURL()
